Support 64-bit payload pointers and add release-profile optimizations
- Change all payload pointers from u32 to u64:
  - LookupEntry.payload_ptr in lookup_table_store (read/write, append/get)
  - PayloadStore append()/get() signatures and return values
  - Builder append_batch buffer and related types
- Update README to document 8-byte little-endian lookup pointers
- Enable aggressive release-profile settings in Cargo.toml:
    opt-level = 3, lto = true, debug = false, panic = "abort", overflow-checks = false

+++ src/builder.rs
@@ -135,7 +135,7 @@ impl IndexedBuilder {
         Ok(())
     }
 
-    pub fn append_batch(&mut self, entries: &mut Vec<(Vec<u8>, u32)>) -> Result<()> {
+    pub fn append_batch(&mut self, entries: &mut Vec<(Vec<u8>, u64)>) -> Result<()> {
         if entries.is_empty() {
             return Ok(());
         }
@@ -180,7 +180,7 @@ where
 {
     base: PathBuf,
     payload_store: PayloadStoreBuilder<V>,
-    buffer: Vec<(Vec<u8>, u32)>,
+    buffer: Vec<(Vec<u8>, u64)>,
     idx_builder: Option<IndexedBuilder>,
     /// Optional callback for each key merged during final index write.
     write_cb: Box<dyn FnMut()>,
